+++ mitra_utils/src/lib.rs
@@ -4,6 +4,7 @@ pub mod files;
 pub mod html;
 pub mod id;
 pub mod identicons;
+pub mod languages;
 pub mod markdown;
 pub mod passwords;
 pub mod random;