+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Allow changing default configuration file path at compile time.
 - Create reposts when processing `Announce(Create(Note))` for top-level group posts.
+- Support per-post language selection.
 
 ### Changed